enable OData v2 (#209)

+++ test/bin/testODataV4DirectRemote.sh
@@ -1,21 +1,23 @@
 #!/bin/bash
 
 set -e
-#set -x
+set -x
 
-TEST_DIR=./test/Vega
+ROOT_DIR=`pwd`
 
-cd ${TEST_DIR}
-WORK_DIR=`pwd`
+WORK_DIR=${ROOT_DIR}/tmp/odata-v4-direct-remote
 
-export NODE_PATH=${WORK_DIR}/node_modules
+test -d $WORK_DIR && rm -r -f $WORK_DIR
+test -d $WORK_DIR || mkdir -p $WORK_DIR
 
-#test -d node_modules && rm -r -f node_modules
-#test -f package-lock.json && rm package-lock.json
-#test -f db.sqlite && rm db.sqlite
+cd test/Vega
+
+cp -v -r srv db _i18n .cdsrc.json .gitignore cucumber.yml package.json $WORK_DIR
+
+cd $WORK_DIR
 
 test -d node_modules || npm i
-rm db.sqlite
+test -f db.sqlite && rm db.sqlite
 test -f db.sqlite || npx cds deploy --to sqlite
 
 # start service
@@ -25,7 +27,8 @@ echo "Started process with PID ${PID}"
 
 # test
 set +e
-REMOTE_PORT=7007 npx cucumber-js ../../test/features/odata --tags "not @todo:remote"
+
+REMOTE_PORT=7007 npx cucumber-js ../../test/features/odata --tags "not @todo:remote" --fail-fast
 RC=$?
 
 # clean up
